# backend/tests/fixtures/api_contract.json
{
  "components": {
    "schemas": {
...
      "GameAnalysis": {
        "additionalProperties": false,
        "properties": {
          "best_move": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Move"
          },
          "report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GameReviewReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "score": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Score"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "report",
          "score",
          "best_move"
        ],
        "title": "GameAnalysis",
        "type": "object"
      },
      "GameDetail": {
        "additionalProperties": false,
        "properties": {
          "accuracy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GameAccuracy"
              },
              {
                "type": "null"
              }
            ]
          },
          "black": {
            "title": "Black",
            "type": "string"
          },
          "black_rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Black Rating"
          },
          "frames": {
            "items": {
              "$ref": "#/components/schemas/GameFrame"
            },
            "title": "Frames",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewJob"
              },
              {
                "type": "null"
              }
            ]
          },
          "orientation": {
            "enum": [
              "white",
              "black"
            ],
            "title": "Orientation",
            "type": "string"
          },
          "played_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Played On"
          },
          "rating": {
            "title": "Rating",
            "type": "integer"
          },
          "result": {
            "title": "Result",
            "type": "string"
          },
          "white": {
            "title": "White",
            "type": "string"
          },
          "white_rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "White Rating"
          }
        },
        "required": [
          "id",
          "white",
          "black",
          "played_on",
          "result",
          "orientation",
          "rating",
          "white_rating",
          "black_rating",
          "frames",
          "job",
          "accuracy"
        ],
        "title": "GameDetail",
        "type": "object"
      },
...
      "GameHistory": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GameHistoryItem"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "GameHistory",
        "type": "object"
      },
      "GameHistoryItem": {
        "additionalProperties": false,
        "properties": {
          "accuracy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GameAccuracy"
              },
              {
                "type": "null"
              }
            ]
          },
          "black": {
            "title": "Black",
            "type": "string"
          },
          "black_rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Black Rating"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "learner_color": {
            "enum": [
              "white",
              "black"
            ],
            "title": "Learner Color",
            "type": "string"
          },
          "move_count": {
            "title": "Move Count",
            "type": "integer"
          },
          "played_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Played At"
          },
          "played_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Played On"
          },
          "result": {
            "title": "Result",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "time_control": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Control"
          },
          "time_control_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Control Label"
          },
          "white": {
            "title": "White",
            "type": "string"
          },
          "white_rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "White Rating"
          }
        },
        "required": [
          "id",
          "white",
          "black",
          "white_rating",
          "black_rating",
          "learner_color",
          "played_on",
          "played_at",
          "result",
          "time_control",
          "time_control_label",
          "move_count",
          "status",
          "accuracy"
        ],
        "title": "GameHistoryItem",
        "type": "object"
      },
...
      "GamePosition": {
        "additionalProperties": false,
        "properties": {
          "fen": {
            "title": "Fen",
            "type": "string"
          },
          "legal_moves": {
            "items": {
              "$ref": "#/components/schemas/LegalMove"
            },
            "title": "Legal Moves",
            "type": "array"
          },
          "result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "san": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "San"
          },
          "termination": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Termination"
          },
          "turn": {
            "enum": [
              "white",
              "black"
            ],
            "title": "Turn",
            "type": "string"
          }
        },
        "required": [
          "fen",
          "turn",
          "legal_moves",
          "result",
          "termination",
          "san"
        ],
        "title": "GamePosition",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobStarted": {
        "additionalProperties": false,
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "title": "JobStarted",
        "type": "object"
      },
...
      "ReviewProgress": {
        "additionalProperties": false,
        "properties": {
          "accuracy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GameAccuracy"
              },
              {
                "type": "null"
              }
            ]
          },
          "job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewJob"
              },
              {
                "type": "null"
              }
            ]
          },
          "moves": {
            "items": {
              "$ref": "#/components/schemas/ReviewedMove"
            },
            "title": "Moves",
            "type": "array"
          }
        },
        "required": [
          "job",
          "moves",
          "accuracy"
        ],
        "title": "ReviewProgress",
        "type": "object"
      },
...
      "ReviewRequest": {
        "properties": {
          "rating": {
            "anyOf": [
              {
                "maximum": 3000.0,
                "minimum": 400.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          }
        },
        "title": "ReviewRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VariationRequest": {
        "properties": {
          "moves": {
            "items": {
              "type": "string"
            },
            "maxItems": 128,
            "title": "Moves",
            "type": "array"
          },
          "ply": {
            "default": 0,
            "minimum": 0.0,
            "title": "Ply",
            "type": "integer"
          }
        },
        "title": "VariationRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Fieldwork API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/games": {
      "get": {
        "operationId": "games_api_games_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameHistory"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Games"
      }
    },
    "/api/games/{game_id}": {
      "get": {
        "operationId": "game_detail_api_games__game_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Game Detail"
      }
    },
    "/api/games/{game_id}/analyze": {
      "post": {
        "operationId": "analyze_variation_api_games__game_id__analyze_post",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VariationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameAnalysis"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze Variation"
      }
    },
    "/api/games/{game_id}/position": {
      "post": {
        "operationId": "variation_position_api_games__game_id__position_post",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VariationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GamePosition"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Variation Position"
      }
    },
    "/api/games/{game_id}/review": {
      "get": {
        "operationId": "review_progress_api_games__game_id__review_get",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewProgress"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Progress"
      },
      "post": {
        "operationId": "begin_review_api_games__game_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStarted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Begin Review"
      }
    },
    "/api/games/{game_id}/train": {
      "post": {
        "operationId": "train_game_api_games__game_id__train_post",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStarted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Train Game"
      }
    },
...
  }
}